Made a really bad agent, and changed datatypes used
Changed datatypes, since uint and int lower than int32 turned out to be incompatible with other parts of tensorflow, now everything is int64 or int32.

I do not yet fully understand what is going on with the agent, but what confuses me most is why it is performing so much worse than the policy making random moves, it feels as if though that should be a baseline performence level.